backend/requirements.txt
fastapi>=0.68.0
uvicorn[standard]>=0.15.0
python-multipart>=0.0.5
pymongo>=4.0.0
motor>=3.0.0
pydantic>=1.8.0
bcrypt>=4.0.0
PyJWT>=2.8.0
python-dotenv>=1.0.0